Datatypes package renamed to match package structure convention (#204)
* datatypes package renamed to model

* Formatting
