+++ web-ui/src/main/java/org/archcnl/domain/input/io/ArchRulesExporter.java
@@ -2,7 +2,7 @@ package org.archcnl.domain.input.io;
 
 import java.io.File;
 import java.io.IOException;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
 
 public interface ArchRulesExporter {
