+++ web-ui/src/main/java/org/archcnl/domain/input/io/ArchRulesFromAdocReader.java
@@ -10,17 +10,6 @@ import java.util.regex.Pattern;
 import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
-import org.archcnl.domain.input.datatypes.architecturerules.ArchitectureRule;
-import org.archcnl.domain.input.datatypes.mappings.AndTriplets;
-import org.archcnl.domain.input.datatypes.mappings.ConceptMapping;
-import org.archcnl.domain.input.datatypes.mappings.CustomConcept;
-import org.archcnl.domain.input.datatypes.mappings.CustomRelation;
-import org.archcnl.domain.input.datatypes.mappings.ObjectType;
-import org.archcnl.domain.input.datatypes.mappings.Relation;
-import org.archcnl.domain.input.datatypes.mappings.RelationMapping;
-import org.archcnl.domain.input.datatypes.mappings.Triplet;
-import org.archcnl.domain.input.datatypes.mappings.Variable;
 import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
 import org.archcnl.domain.input.exceptions.NoArchitectureRuleException;
 import org.archcnl.domain.input.exceptions.NoMappingException;
@@ -29,6 +18,17 @@ import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.exceptions.UnrelatedMappingException;
 import org.archcnl.domain.input.exceptions.UnsupportedObjectTypeInTriplet;
 import org.archcnl.domain.input.exceptions.VariableAlreadyExistsException;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.input.model.mappings.AndTriplets;
+import org.archcnl.domain.input.model.mappings.ConceptMapping;
+import org.archcnl.domain.input.model.mappings.CustomConcept;
+import org.archcnl.domain.input.model.mappings.CustomRelation;
+import org.archcnl.domain.input.model.mappings.ObjectType;
+import org.archcnl.domain.input.model.mappings.Relation;
+import org.archcnl.domain.input.model.mappings.RelationMapping;
+import org.archcnl.domain.input.model.mappings.Triplet;
+import org.archcnl.domain.input.model.mappings.Variable;
 
 public class ArchRulesFromAdocReader implements ArchRulesImporter {
