+++ web-ui/src/main/java/org/archcnl/domain/input/io/ArchRulesToAdocWriter.java
@@ -6,11 +6,11 @@ import java.nio.charset.StandardCharsets;
 import java.util.LinkedList;
 import java.util.List;
 import org.apache.commons.io.FileUtils;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
-import org.archcnl.domain.input.datatypes.architecturerules.ArchitectureRule;
-import org.archcnl.domain.input.datatypes.mappings.CustomConcept;
-import org.archcnl.domain.input.datatypes.mappings.CustomRelation;
-import org.archcnl.domain.input.datatypes.mappings.Mapping;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.input.model.mappings.CustomConcept;
+import org.archcnl.domain.input.model.mappings.CustomRelation;
+import org.archcnl.domain.input.model.mappings.Mapping;
 
 public class ArchRulesToAdocWriter implements ArchRulesExporter {
