+++ web-ui/src/main/java/org/archcnl/domain/input/model/architecturerules/ArchitectureRule.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.datatypes.architecturerules;
+package org.archcnl.domain.input.model.architecturerules;
 
 import java.util.Objects;
