+++ web-ui/src/main/java/org/archcnl/domain/input/model/mappings/Concept.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.datatypes.mappings;
+package org.archcnl.domain.input.model.mappings;
 
 import java.util.Objects;
