+++ web-ui/src/main/java/org/archcnl/domain/input/model/mappings/TypeRelation.java
@@ -1,9 +1,9 @@
-package org.archcnl.domain.input.datatypes.mappings;
+package org.archcnl.domain.input.model.mappings;
 
 import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
 
 public class TypeRelation extends Relation {
