+++ web-ui/src/test/java/org/archcnl/domain/input/TestUtils.java
@@ -4,16 +4,6 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
-import org.archcnl.domain.input.datatypes.architecturerules.ArchitectureRule;
-import org.archcnl.domain.input.datatypes.mappings.AndTriplets;
-import org.archcnl.domain.input.datatypes.mappings.ConceptMapping;
-import org.archcnl.domain.input.datatypes.mappings.CustomConcept;
-import org.archcnl.domain.input.datatypes.mappings.CustomRelation;
-import org.archcnl.domain.input.datatypes.mappings.RelationMapping;
-import org.archcnl.domain.input.datatypes.mappings.StringValue;
-import org.archcnl.domain.input.datatypes.mappings.Triplet;
-import org.archcnl.domain.input.datatypes.mappings.Variable;
 import org.archcnl.domain.input.exceptions.ConceptAlreadyExistsException;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
@@ -22,6 +12,16 @@ import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.exceptions.UnrelatedMappingException;
 import org.archcnl.domain.input.exceptions.UnsupportedObjectTypeInTriplet;
 import org.archcnl.domain.input.exceptions.VariableAlreadyExistsException;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.input.model.mappings.AndTriplets;
+import org.archcnl.domain.input.model.mappings.ConceptMapping;
+import org.archcnl.domain.input.model.mappings.CustomConcept;
+import org.archcnl.domain.input.model.mappings.CustomRelation;
+import org.archcnl.domain.input.model.mappings.RelationMapping;
+import org.archcnl.domain.input.model.mappings.StringValue;
+import org.archcnl.domain.input.model.mappings.Triplet;
+import org.archcnl.domain.input.model.mappings.Variable;
 
 public class TestUtils {
