+++ web-ui/src/test/java/org/archcnl/domain/input/io/ArchRulesFromAdocReaderTest.java
@@ -6,10 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.io.File;
 import java.io.IOException;
 import org.archcnl.domain.input.TestUtils;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
-import org.archcnl.domain.input.datatypes.architecturerules.ArchitectureRule;
-import org.archcnl.domain.input.datatypes.mappings.Concept;
-import org.archcnl.domain.input.datatypes.mappings.Relation;
 import org.archcnl.domain.input.exceptions.ConceptAlreadyExistsException;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
@@ -18,6 +14,10 @@ import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.exceptions.UnrelatedMappingException;
 import org.archcnl.domain.input.exceptions.UnsupportedObjectTypeInTriplet;
 import org.archcnl.domain.input.exceptions.VariableAlreadyExistsException;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
+import org.archcnl.domain.input.model.mappings.Concept;
+import org.archcnl.domain.input.model.mappings.Relation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
