+++ web-ui/src/test/java/org/archcnl/domain/input/mappings/ConceptManagerTest.java
@@ -7,14 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.LinkedList;
 import java.util.List;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
-import org.archcnl.domain.input.datatypes.mappings.AndTriplets;
-import org.archcnl.domain.input.datatypes.mappings.ConceptManager;
-import org.archcnl.domain.input.datatypes.mappings.ConceptMapping;
-import org.archcnl.domain.input.datatypes.mappings.CustomConcept;
-import org.archcnl.domain.input.datatypes.mappings.DefaultConcept;
-import org.archcnl.domain.input.datatypes.mappings.Triplet;
-import org.archcnl.domain.input.datatypes.mappings.Variable;
 import org.archcnl.domain.input.exceptions.ConceptAlreadyExistsException;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
@@ -22,6 +14,14 @@ import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.exceptions.UnrelatedMappingException;
 import org.archcnl.domain.input.exceptions.UnsupportedObjectTypeInTriplet;
 import org.archcnl.domain.input.exceptions.VariableAlreadyExistsException;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.mappings.AndTriplets;
+import org.archcnl.domain.input.model.mappings.ConceptManager;
+import org.archcnl.domain.input.model.mappings.ConceptMapping;
+import org.archcnl.domain.input.model.mappings.CustomConcept;
+import org.archcnl.domain.input.model.mappings.DefaultConcept;
+import org.archcnl.domain.input.model.mappings.Triplet;
+import org.archcnl.domain.input.model.mappings.Variable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
