+++ web-ui/src/test/java/org/archcnl/domain/input/mappings/ConceptMappingTest.java
@@ -6,8 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.LinkedList;
 import java.util.List;
 import org.archcnl.domain.input.TestUtils;
-import org.archcnl.domain.input.datatypes.mappings.ConceptMapping;
-import org.archcnl.domain.input.datatypes.mappings.CustomConcept;
 import org.archcnl.domain.input.exceptions.ConceptAlreadyExistsException;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
@@ -16,6 +14,8 @@ import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.exceptions.UnrelatedMappingException;
 import org.archcnl.domain.input.exceptions.UnsupportedObjectTypeInTriplet;
 import org.archcnl.domain.input.exceptions.VariableAlreadyExistsException;
+import org.archcnl.domain.input.model.mappings.ConceptMapping;
+import org.archcnl.domain.input.model.mappings.CustomConcept;
 import org.junit.jupiter.api.Test;
 
 class ConceptMappingTest {
