+++ web-ui/src/test/java/org/archcnl/domain/input/mappings/RelationManagerTest.java
@@ -7,17 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.LinkedList;
 import java.util.List;
-import org.archcnl.domain.input.datatypes.RulesConceptsAndRelations;
-import org.archcnl.domain.input.datatypes.mappings.AndTriplets;
-import org.archcnl.domain.input.datatypes.mappings.ConceptManager;
-import org.archcnl.domain.input.datatypes.mappings.CustomRelation;
-import org.archcnl.domain.input.datatypes.mappings.DefaultRelation;
-import org.archcnl.domain.input.datatypes.mappings.RelationManager;
-import org.archcnl.domain.input.datatypes.mappings.RelationMapping;
-import org.archcnl.domain.input.datatypes.mappings.SpecialRelation;
-import org.archcnl.domain.input.datatypes.mappings.Triplet;
-import org.archcnl.domain.input.datatypes.mappings.TypeRelation;
-import org.archcnl.domain.input.datatypes.mappings.Variable;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
 import org.archcnl.domain.input.exceptions.RelationAlreadyExistsException;
@@ -25,6 +14,17 @@ import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.exceptions.UnrelatedMappingException;
 import org.archcnl.domain.input.exceptions.UnsupportedObjectTypeInTriplet;
 import org.archcnl.domain.input.exceptions.VariableAlreadyExistsException;
+import org.archcnl.domain.input.model.RulesConceptsAndRelations;
+import org.archcnl.domain.input.model.mappings.AndTriplets;
+import org.archcnl.domain.input.model.mappings.ConceptManager;
+import org.archcnl.domain.input.model.mappings.CustomRelation;
+import org.archcnl.domain.input.model.mappings.DefaultRelation;
+import org.archcnl.domain.input.model.mappings.RelationManager;
+import org.archcnl.domain.input.model.mappings.RelationMapping;
+import org.archcnl.domain.input.model.mappings.SpecialRelation;
+import org.archcnl.domain.input.model.mappings.Triplet;
+import org.archcnl.domain.input.model.mappings.TypeRelation;
+import org.archcnl.domain.input.model.mappings.Variable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
