+++ web-ui/src/test/java/org/archcnl/domain/input/mappings/RelationMappingTest.java
@@ -5,8 +5,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.LinkedList;
 import java.util.List;
 import org.archcnl.domain.input.TestUtils;
-import org.archcnl.domain.input.datatypes.mappings.CustomRelation;
-import org.archcnl.domain.input.datatypes.mappings.RelationMapping;
 import org.archcnl.domain.input.exceptions.ConceptAlreadyExistsException;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.InvalidVariableNameException;
@@ -15,6 +13,8 @@ import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.exceptions.UnrelatedMappingException;
 import org.archcnl.domain.input.exceptions.UnsupportedObjectTypeInTriplet;
 import org.archcnl.domain.input.exceptions.VariableAlreadyExistsException;
+import org.archcnl.domain.input.model.mappings.CustomRelation;
+import org.archcnl.domain.input.model.mappings.RelationMapping;
 import org.junit.jupiter.api.Test;
 
 class RelationMappingTest {
